requirements.txt
torch
einops
pyyaml
